Load the configuration so that example config file works
The example config file (`config/mangoDB`) has multiple connections set as the top level keys in the configuration array - as common for database configuration in Kohana modules. But the `MangoDB::instance()` method expects the connection details to be the top level keys in a file named after the connection name.

I changed the configuration loading to work as the example file expects and as common practice in Kohana.